requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.1
requests>=2.31.0
twilio>=9.0.0
django-parler>=2.3
django-rosetta>=0.10.0
Pillow>=10.2.0
stripe>=8.0.0
